+++ src/callers/tx.ts
@@ -21,8 +21,11 @@ const createWhereFilter = (sender: PublicKey | null, operator?: string) => {
   const exp = { signer_infos: {}, block: {} }
 
   const escapedString = sender?.key.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
+
   const regex = new RegExp(escapedString!)
-  const regexString = regex.toString().replace('/', '').slice(0, -1)
+
+  // Remove the first and last character of the regex string to remove the / at the start and end
+  const regexString = regex.toString().substring(1, regex.toString().length - 1)
 
   if (sender) {
     exp.signer_infos = {