+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+### Fixed
+- `tx` caller `where` filter
+
 ## [1.2.0] - 2024-03-19
 ### Added
 - `Text` Proposal type